refactor: separate storage_service and meta_event_service
